@@ -49,11 +49,11 @@ class SomeEvent
     public string Name {get; set;}
 }";
         DiagnosticResult[] expected =
-        {
+        [
             Diagnostic(DescriptorRules.InvalidIdentity)
                 .WithSpan(4, 2, 4, 15)
                 .WithArguments("EventType", "eventTypeId", "\"\"")
-        };
+        ];
 
         await VerifyAnalyzerAsync(test, expected);
     }
@@ -68,11 +68,11 @@ class SomeEvent
     public string Name {get; set;}
 }";
         DiagnosticResult[] expected =
-        {
+        [
             Diagnostic(DescriptorRules.InvalidIdentity)
                 .WithSpan(2, 2, 2, 35)
-                .WithArguments("Dolittle.SDK.Events.EventType", "eventTypeId", "\"\""),
-        };
+                .WithArguments("Dolittle.SDK.Events.EventType", "eventTypeId", "\"\"")
+        ];
 
         await VerifyAnalyzerAsync(test, expected);
     }
@@ -113,11 +113,11 @@ class SomeEvent
     public string Name {get; set;}
 }";
         DiagnosticResult[] expected =
-        {
+        [
             Diagnostic(DescriptorRules.InvalidIdentity)
                 .WithSpan(2, 2, 2, 80)
                 .WithArguments("Dolittle.SDK.Events.EventType", "eventTypeId", "\"c6f87322-be67-4aaf\"")
-        };
+        ];
 
         await VerifyAnalyzerAsync(test, expected);
     }
@@ -145,11 +145,11 @@ class SomeEvent
     public string Name {get; set;}
 }";
         DiagnosticResult[] expected =
-        {
+        [
             Diagnostic(DescriptorRules.InvalidIdentity)
                 .WithSpan(2, 2, 2, 47)
-                .WithArguments("Dolittle.SDK.Events.Handling.EventHandler", "eventHandlerId", "\"\""),
-        };
+                .WithArguments("Dolittle.SDK.Events.Handling.EventHandler", "eventHandlerId", "\"\"")
+        ];
 
         await VerifyAnalyzerAsync(test, expected);
     }
